feat(cognito): support `emailVerified` for `AttributeMapping` interface (#31632)
### Issue #30467 

Closes #30467

### Reason for this change

For custom OpenId providers, there is no way to automatically validate email upon sign-in.
Therefore, we would like to add the `email_verified` attribute to attribute mapping, but it is not present in the member  definition of `AttributeMapping` interface., so we have added it in this PR.

### Description of changes


Added `emailVerified` attribute to `AttributeMapping` interface.

### Description of how you validated changes


Added the `email_verified` assertion to both unit and integration tests.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*