fix(scheduler-alpha): remove `targetOverrides` prop from Schedule (#31799)
### Issue # (if applicable)

N/A

### Reason for this change

The `targetOverrides` prop is adding confusion to the API as it override what is set on the target. A better way to reuse target props would be:
```ts
const targetBaseProps = { input: ...someInput... }

const schedule1 = new Schedule(this, 'Schedule', {
    scheduleExpression: ScheduleExpression.cron({ day: '20' }),
    target: new targets.LambdaInvoke(props.func, {
        ...targetBaseProps,
        // override whatever or not to override
        }),
    }),
});
```

### Description of changes

Removed `targetOverrides` prop and the related tests. Also updated the integ test which was using the prop.

### Description of how you validated changes

Removing a prop so n/a.

### Checklist
- [X] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*

----

BREAKING CHANGE: This PR removes the `targetOverrides` prop from the `Schedule` construct.