fix(cli): deploy-role is not authorized to perform DescribeStackResources (#31878)
The cross-account asset uploading detection check required that the `deploy-role` could call `DescribeStackResources` -- which it can't.

Instead, rely on parsing the `Outputs` of `DescribeStacks`. This is equivalent for the built-in stack, and relies on stack customizers to have removed the Output or put a dummy value there that does not look like a stack name (like `''`, `'-'` or `'*'`). It's not *as* good, but probably good enough.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*