chore(cli): integ tests don't buffer shell output (#31903)
What should happen is that the output of shell commands gets buffered, and only printed if the test running them fails.

In practice, we see the output being printed directly.

The buffering code is a bit confusing, and we don't exactly understand why it's not working. Try to simplify the code a bit and remove mutable variable manipulation, to address the above.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*