refactor: gate access to environment SDK behind new class
Previously there were methods on the `Deployments` class that made it
possible to directly get an SDK from the `SdkProvider` for a particular
environment. Calling these methods made it possible to get an SDK
without thinking of assuming roles to go into a different account.

This PR introduces a new class, `EnvironmentAccess`, with a couple of
public methods that are the only ones allowed to obtain SDKs with
credentials. It has the methods:

- accessStackForStackOperations(stack)
- accessStackForLookup(stack)
- accessStackForReading(stack)

These will always respect the role information on the stack.

Ideally there would have been similar methods for assets as well, but
the `cdk-assets` library is entirely handling asset roles itself, and
it's not in the scope of this PR to change that. That keeps on using a
plain `SdkProvider`. Hotswap deployments will also just use CLI
credentials and not assume role, so that also keeps on using an
`SdkProvider`.

All other uses have moved to `EnvironmentAccess`.