chore(ec2): add missing interface VPC endpoint (#31907)
### Issue # (if applicable)

N/A

### Reason for this change

The `InterfaceVpcEndpointAwsService` class was missing the AWS Price List service.

### Description of changes

Added `InterfaceVpcEndpointAwsService.PRICING_API` to the enum, connected to the service `'pricing.api'` in accordance with [documentation](https://docs.aws.amazon.com/vpc/latest/privatelink/aws-services-privatelink-support.html).

### Description of how you validated changes

Verified the extension via the documentation, and confirmed the endpoint exists in the console: 
![image](https://github.com/user-attachments/assets/4a85b146-fd97-4ec7-94a0-0f7d6551e6f1)

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*