feat(cli): automatically roll back stacks if necessary
If a user is deploying with `--no-rollback`, and the stack contains
replacements (or the `--no-rollback` flag is dropped), then a rollback
needs to be performed before a regular deployment can happen again.

In this PR, we add a prompt where we ask the user to confirm that
they are okay with performing a rollback and then a normal deployment.

Closes #30546.