chore(cli): add image to gc docs and document race condition (#31925)
This might be contentious given that we don't do photos in most places but afaik we do not replicate the CLI docs on the CDK docs site so it may be safe (as it renders correctly on github).

Documents garbage collection algorithm via diagram and details potentail race condition with `REVIEW_IN_PROGRESS` stacks.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*