fix(cli): asset uploads fail if Object Lock is enabled on access bucket
Object Lock requires passing an object checksum. By default, SDKv2
only calculates MD5 checksums.

We used to turn off checksums altogether and rely on SigV4 checksums to
produce a workable setup for both FIPS and non-FIPS users, but in case
of Object Lock this doesn't work: we must definitely have an S3 content
checksum, and the the SigV4 checksum alone is not good enough.

Since SDKv2 only supports MD5 checksums, we now only disable checksums
for FIPS environments.

The unfortunate result is that Object Lock will not work in a FIPS
environment, but there's no way around that for now.

When we migrate to SDKv3, which can be configured to checksum using
SHA256, Object Lock + FIPS will work again.