fix(location): remove base class from PlaceIndex class (#31287)
### Issue # (if applicable)

None

### Reason for this change
I implemented the location resources in #30711 and #30682. 

During the review, it was suggested that it would be better not to have a base class.
As a result, the `GeofenceCollection` class and `RouteCalculator` class do not have a base class. 

https://github.com/aws/aws-cdk/pull/30711#discussion_r1734937879

On the other hand, the `PlaceIndex `class has already been implemented with a base class, which is inconsistent with the current approach.



### Description of changes
Remove a unnecessary base class from `PlaceIndex` class.



### Description of how you validated changes
Re-run unit tests and integ tests.


### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*