chore: fix C# build (#31946)
Currently, the generated C# code doesn't compile with this error:

```
error CS0115: 'VpcV2Base._Proxy.Ipv4IpamProvisionedCidrs': no suitable method found to override
```

The reason is that the abstract base class that implements an interface doesn't have an abstract field for that interface. 

Then when a subclass does implement it, it gets rendered to an `override` in C#, but there is no method to override.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*