feat(bootstrap): delete noncurrent versions after 30 days (#31949)
CDK assets in the bootstrap bucket are content-addressed and immutable so we never naturally create noncurrent versions.
However, with the introduction of the `cdk gc` command, we now have the capacity to delete unused objects in the bucket. Because the bucket is versioned by default, the delete command [actually](https://docs.aws.amazon.com/AmazonS3/latest/userguide/DeletingObjectVersions.html) just replaces the object with a new dummy version (thus the deleted object becomes noncurrent). 

Now that noncurrent objects _can_ happen, if one utilizes `cdk gc`, we are updating the bootstrap template to retain them for 30 days rather than 365 days.

Update to bootstrap version 24 to use this new lifecycle policy in conjunction with `cdk gc`. 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*