chore: automatically age out old TypeScript versions from testing
In our integ tests, we were testing TypeScript versions from 3.9 up to
the latest version, which recently broke because of a change to
modernize the init templates.

We should up this range to a recent version; the common support
lifetime of a TypeScript version is the one on `DefinitelyTyped`, the
types registry for TypeScript packages. They only target TS versions
less than 2 years old, which at the time of this PR is 4.9 and higher.

Encode that policy into code automatically, so that we don't have to
manually keep this minimum version up-to-date. This currently ages
out the following versions:

```
  ✕ typescript 3.9 init app
  ✕ typescript 4.0 init app
  ✕ typescript 4.1 init app
  ✕ typescript 4.2 init app
  ✕ typescript 4.3 init app
  ✕ typescript 4.4 init app
  ✕ typescript 4.5 init app
  ✕ typescript 4.6 init app
  ✕ typescript 4.7 init app
  ✕ typescript 4.8 init app
------------------------------8< cut line -----
  ✕ typescript 4.9 init app
  ✓ typescript 5.0 init app
  ✓ typescript 5.1 init app
  ✓ typescript 5.2 init app
  ✓ typescript 5.3 init app
  ✓ typescript 5.4 init app
  ✓ typescript 5.5 init app
  ✓ typescript 5.6 init app
```

Unfortunately not enough to save the TypeScript template modernization
change entirely... but at least it's an improvement.