chore(pipes-targets): remove dependency with pipes-sources (#31986)
Update the integration test cases for both pipes-sources and pipes-targets modules to add comments explained why we use TestTarget, and TestSource instead of using real classes, and this to avoid the circular dependencies, as we need to depend on each module on the other so we can implement these test cases.

These Test classes should be replaced before graduating the pipes alpha modules.