fix: deploy-time stack tags cause synthesis to fail
In https://github.com/aws/aws-cdk/pull/31457, we introduced a change
that made synthesis fail if one of the stack tags was a deploy-time
value. Since stack tags are assigned outside a CloudFormation context,
deploy-time values cannot be evaluated, so the stack ends up with
a tag like `{ Key: "my-tag", Value: "${Token[1234]}" }`, which is
probably not what is intended.

Worse, those tags are automatically propagated to all resources in the
stack by CloudFormation, and some may validate the tag value and find
that `$` or any of the other characters are not valid tag values.

The intent was that customers would be alerted to these kinds of
mistakes and apply their tags to resources, or skip stacks
when applying tags to large scopes:

```ts
Tags.of(this).add('my-tag', Fn.importValue('SomeExport'), {
  excludeResourceTypes: ['aws:cdk:stack'],
});
```

The previous change was a bit drastic in its attempts.  In this one we
ignore the unresolved tags and add a warning instead. That way,
synthesis still succeeds.