chore: make `@aws-cdk/yargs-gen` a devDependency
`yargs-gen` used to be a runtime dependency. The only reason seemed to
be that there was a factory class for `DynamicResult` types in
`yargs-gen`, called `DynamicValue`.

In this PR, move the factory to the only location where it is used
in the CLI itself, and turn the `import` of `yargs-gen` into an
`import type`, which does not imply any runtime dependency.