fix(scheduler-targets-alpha): imported target resources as schedule target throws synth error (#32105)
### Issue # (if applicable)

Tracking #31785 

### Reason for this change

Similar to what was done in this PR for the Lambda Invoke target: https://github.com/aws/aws-cdk/pull/31837. 
Reasoning is explained in this [comment](https://github.com/aws/aws-cdk/pull/29615#issuecomment-2417858246):
> At synth time, there are cases where CDK does not know about the actual environment (e.g. when the environment is using Tokens). In other words, the environment check works in some scenarios and not others. This creates more confusion than benefit.
IF cross env target support is added in the future, this check will become a blocker for CDK customers to set up cross env target.


### Description of changes

Remove synth-time error that checks for same account and region for all targets to allow use of imported resources as target.

### Description of how you validated changes

All tests are passing.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*