fix(eks): eks.AccessEntryType.STANDARD maps to a wrong value (#31147)
### Reason for this change

Tried making use of the `eks.AccessEntry` construct, with the `eks.AccessEntryType.STANDARD` as accessEntryType, but the stack deployment failed with the following error: `AccessEntry type must be one of [EC2_LINUX, EC2_WINDOWS, FARGATE_LINUX, STANDARD]`.

Seems like 'Standard' is not alright, suggested also by `eksctl` [docs](https://eksctl.io/usage/access-entries/#iam-entities).

If this is the case, documentation in the following places also needs to be corrected:

- https://docs.aws.amazon.com/eks/latest/APIReference/API_CreateAccessEntry.html
- https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-eks-accessentry.html
- https://docs.aws.amazon.com/cdk/api/v2/docs/aws-cdk-lib.aws_eks.CfnAccessEntry.html#construct-props

### Description of changes



### Description of how you validated changes



### Checklist
- [X] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*