chore(cli): update tests for future monorepo (#32185)
A number of our tests depend on the version of the `cloud-assembly-schema`, and some of our tests use `aws-cdk-lib` to generate Cloud Assemblies which has its own copy of `cloud-assembly-schema` built in.

Because `cloud-assembly-schema` currently takes its schema version from its `package.json` version, and because the version of `cloud-assembly-schema/package.json` will be `0.0.0` once it has been moved into our new monorepo, these tests need to be changed to deal with this situation.

The solution we're currently taking is to rewrite `manifest.json` and change the `version` in there to a different one.

This PR also renames `expect(...).toBeCalledWith()` to `expect(...).toHaveBeenCalledWith()` because the jest linter requires is.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*